Added endpoint for uploading files in AppFileController
Decided to split functionality of working with files into two controllers:
AppFileController will work on everything that related to AppFile as a database entity. It won't send files, only references to them in Azure Blob Storage and SQL Server.
FileController will send actual data of the file.

Added MediaService, but probably gonna change the name and directory to MediaUtilities. This class perform operations on media files like images, videos, audios, maybe in the future I am gonna learn how to work on pdfs with some kind of a library but it's not gonna be anywhen soon.

All responses in all microservices are gonna send the same response class but different requests.